:ok_hand: Update code due to static analysis review
Once the CI is not running well due to Compose and new Gradle issues,
the static analysis were executed locally and the code is now fresh
again.
